Added code for function isPrime
The function isPrime was filled in. If the number was 0 or 1 it fails per convention of prime numbers. All other numbers are tested to see if they ever can be divided by a number smaller than them and have no remainder; if they do have no remainder then they can't be prime.